requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.16.0
jupyter-client==6.1.7
jupyter-core==4.6.3
mibian==0.1.3
numpy==1.19.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pyzmq==19.0.2
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
yahoo-earnings-calendar==0.6.0
